backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
docopt==0.6.2
drf-yasg==1.20.0
gensim==3.6.0
idna==3.3
importlib-resources==5.6.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
MarkupSafe==2.1.1
mongoengine==0.24.1
neo4j==4.4.2
nltk==3.7
numpy==1.22.3
packaging==21.3
pipreqs==0.4.11
pymongo==3.12.3
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
rake-nltk==1.0.6
regex==2022.3.15
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
sqlparse==0.2.4
swagger-spec-validator==2.7.4
textdistance==4.2.2
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
yarg==0.1.9
zipp==3.7.0
